+++ src/chat/chat-action.tsx
@@ -1,5 +1,7 @@
 import { h, Component } from 'preact';
 import {botman} from './botman';
+import { IAction, IMessage } from '../typings';
+
 
 export default class ChatAction extends Component<IChatActionProps, any> {
 
@@ -23,50 +25,6 @@ export default class ChatAction extends Component<IChatActionProps, any> {
             });
         }, null);
     }
-}
-
-export enum IMessageTypes {
-    'text'
-}
-
-export interface IMessage {
-    text: string,
-    type: IMessageTypes,
-    from: string,
-    time: string,
-    actions: IAction[],
-    buttons: IButton[],
-    attachment: IAttachment,
-    globalButtons: IButton[],
-    additionalParameters: String[],
-    elements: IElement[],
-}
-
-export interface IAttachment {
-    url: string,
-    type: string,
-}
-
-export interface IElement {
-    title : string,
-    image_url: string,
-    item_url: string,
-    subtitle: string,
-    default_action: IAction,
-    buttons: IButton[],
-}
-
-
-export interface IButton {
-    type: string,
-    title: string,
-    url: string,
-    payload: string,
-}
-
-export interface IAction {
-    text: string,
-    value: string,
 }
 
 interface IChatActionProps {
