+++ tsconfig.json
@@ -14,6 +14,6 @@
 		"lib": ["dom", "es5", "scripthost", "es2015.promise"]
 	},
 	"include": [
-		"./src/"
+		"./src/**/*"
 	]
 }
